Refatoracao feita para permitir o uso de blueprints e modularizar a aplicacao
