+++ app/templates/auth/register.html
@@ -28,7 +28,7 @@
     <div class="card-body register-card-body">
       <p class="login-box-msg">Register a new membership</p>
 
-      <form action="{{ url_for('register') }}" method="post">
+      <form action="{{ url_for('auth.register') }}" method="post">
         <div class="input-group mb-3">
           <input type="text" class="form-control" placeholder="Full name" name="name">
           <div class="input-group-append">
@@ -86,7 +86,7 @@
         </div>
       </form>
 
-      <a href="{{ url_for('login') }}" class="text-center">I already have a membership</a>
+      <a href="{{ url_for('auth.login') }}" class="text-center">I already have a membership</a>
     </div>
     <!-- /.form-box -->
   </div><!-- /.card -->
