+++ app/templates/base.html
@@ -138,7 +138,7 @@ scratch. This page gets rid of all links and provides the needed markup only.
       </li>
       <!-- Off -->
       <li class="nav-item">
-        <a class="nav-link"href="{{ url_for('logout') }}">
+        <a class="nav-link"href="{{ url_for('auth.logout') }}">
           Sair
         </a>
       </li>
@@ -164,7 +164,7 @@ scratch. This page gets rid of all links and provides the needed markup only.
           <img src="{{ url_for('static', filename='dist/img/anonymous2.png') }}" class="img-circle elevation-2" alt="User Image">
         </div>
         <div class="info">
-          <a href="{{ url_for('profile') }}" class="d-block">{{usuario}}</a>
+          <a href="{{ url_for('dashboard.profile') }}" class="d-block">{{usuario}}</a>
         </div>
       </div>
