+++ app/templates/dashboard/profile.html
@@ -13,7 +13,7 @@
               <div class="card card-primary card-outline">
                 <div class="card-body box-profile">
                   <div class="text-center">
-                    <a href="{{ url_for('update', id=id)}}"><img class="profile-user-img img-fluid img-circle" src="{{ url_for('static', filename='dist/img/anonymous2.png') }}" alt="User profile picture"></a>
+                    <a href="{{ url_for('user.update', id=id)}}"><img class="profile-user-img img-fluid img-circle" src="{{ url_for('static', filename='dist/img/anonymous2.png') }}" alt="User profile picture"></a>
                   </div>
   
                   <h3 class="profile-username text-center">{{usuario}} </h3>
